refactor(frontend): split playtest particle runtime
